+++ PlayerClient/SmartServerTalker.cs
@@ -45,7 +45,7 @@ namespace PlayerClient
 				var direction = ai.MakeStep();
 				Json.Write(socket, new Move
 				{
-					Direction = direction
+					Direction = (int)direction
 				});
 				var moveResultInfo = Json.Read<MoveResultInfo>(socket);
 				if (moveResultInfo.Result == 0)
